+++ amazonorders/selectors.py
@@ -79,10 +79,14 @@ class Selectors:
                                      "div.order"]
     ORDER_HISTORY_COUNT_SELECTOR = ".js-yo-container span.num-orders"
     ORDER_DETAILS_ENTITY_SELECTOR = ["div#orderDetails",
-                                     "div#ordersContainer"]
+                                     "div#ordersContainer",
+                                     # The Whole Foods Market in-store (FOPO) order details page
+                                     "div#odp-main-section"]
     ITEM_ENTITY_SELECTOR = ["[data-component='purchasedItems'] .a-fixed-left-grid",
                             "div:has(> div.yohtmlc-item)",
-                            ".item-box"]
+                            ".item-box",
+                            # Whole Foods Market in-store (FOPO) order details line items
+                            "div.a-row.a-spacing-base:has(img.ufpo-itemListWidget-image)"]
     SHIPMENT_ENTITY_SELECTOR = ["[data-component='orderCard'] [data-component='shipments'] .a-box",
                                 "div.shipment",
                                 "div.delivery-box"]
@@ -105,7 +109,9 @@ class Selectors:
     ORDER_WHOLE_FOODS = [
         "a[href*='/wholefoodsmarket/receipts/order/']",
         "a[href*='/fopo/order-details']",
-        Selector("div.yohtmlc-shipment-status-primaryText", text_contains="Whole Foods Market")
+        Selector("div.yohtmlc-shipment-status-primaryText", text_contains="Whole Foods Market"),
+        # The item list widget on the Whole Foods in-store (FOPO) order details page
+        "img.ufpo-itemListWidget-image"
     ]
     # Selectors defined here mean the Order will not have parsable totals
     ORDER_SKIP_TOTALS = [
@@ -119,19 +125,24 @@ class Selectors:
     # CSS selectors for Item fields
     #####################################
 
-    FIELD_ITEM_IMG_LINK_SELECTOR = "a img"
+    # The trailing entries (``ufpo``/``a-span10``) match Whole Foods Market in-store (FOPO) line items
+    FIELD_ITEM_IMG_LINK_SELECTOR = ["a img",
+                                    "img.ufpo-itemListWidget-image"]
     FIELD_ITEM_QUANTITY_SELECTOR = [".od-item-view-qty",
                                     "span.item-view-qty",
                                     "span.product-image__qty"]
     FIELD_ITEM_TITLE_SELECTOR = ["[data-component='itemTitle']",
-                                 ".yohtmlc-item a", ".yohtmlc-product-title"]
+                                 ".yohtmlc-item a", ".yohtmlc-product-title",
+                                 "div.a-column.a-span10 > a"]
     FIELD_ITEM_LINK_SELECTOR = ["[data-component='itemTitle'] a",
                                 ".yohtmlc-item a",
                                 "a:has(> .yohtmlc-product-title)",
-                                ".yohtmlc-product-title a"]
+                                ".yohtmlc-product-title a",
+                                "div.a-column.a-span10 > a"]
     FIELD_ITEM_TAG_ITERATOR_SELECTOR = [".yohtmlc-item div"]
     FIELD_ITEM_PRICE_SELECTOR = ["[data-component='unitPrice'] .a-text-price :not(.a-offscreen)",
-                                 ".yohtmlc-item .a-color-price"]
+                                 ".yohtmlc-item .a-color-price",
+                                 "div.a-section.a-text-right span.a-size-small"]
     FIELD_ITEM_SELLER_SELECTOR = ["[data-component='orderedMerchant']"] + FIELD_ITEM_TAG_ITERATOR_SELECTOR
     FIELD_ITEM_RETURN_SELECTOR = ["[data-component='itemReturnEligibility']"] + FIELD_ITEM_TAG_ITERATOR_SELECTOR
 
@@ -142,7 +153,10 @@ class Selectors:
     FIELD_ORDER_DETAILS_LINK_SELECTOR = ["a.yohtmlc-order-details-link",
                                          # Would like to use this or similar, but not yet sure how consistent it is
                                          # ".order-header__header-link-list-item:first-of-type a"
-                                         ]
+                                         # Whole Foods Market receipt orders and in-store (FOPO) purchases link to
+                                         # dedicated detail pages rather than the standard order-details endpoint
+                                         "a[href*='/wholefoodsmarket/receipts/order/']",
+                                         "a[href*='/fopo/order-details']"]
     FIELD_ORDER_NUMBER_SELECTOR = ["[data-component='orderId']",
                                    "[data-component='briefOrderInfo'] div.a-column",
                                    ".order-date-invoice-item :is(bdi, span)[dir='ltr']",
@@ -150,7 +164,12 @@ class Selectors:
                                    ":is(bdi, span)[dir='ltr']"]
     FIELD_ORDER_GRAND_TOTAL_SELECTOR = ["div.yohtmlc-order-total span.value",
                                         "div.order-header div.a-column.a-span2",
-                                        "div.order-header div.a-col-left .a-span9"]
+                                        "div.order-header div.a-col-left .a-span9",
+                                        # Whole Foods Market in-store (FOPO) order details page
+                                        "#wfm-grand-total-amount"]
+    # The "Purchase Summary" amounts on the Whole Foods Market in-store (FOPO) order details page
+    FIELD_ORDER_WHOLE_FOODS_SUBTOTAL_SELECTOR = "#wfm-subtotal-amount"
+    FIELD_ORDER_WHOLE_FOODS_TAX_SELECTOR = "#wfm-tax-total-amount"
     FIELD_ORDER_PLACED_DATE_SELECTOR = ["[data-component='orderDate']",
                                         "span.order-date-invoice-item",
                                         "[data-component='briefOrderInfo'] div.a-column",